requirements.txt
requests
feedparser
beautifulsoup4
pandas
openpyxl
python-dotenv